Handle the case when the bitset size > maxint

+++ src/main/scala/BloomFilterAggregator.scala
@@ -3,7 +3,7 @@ import org.apache.spark.sql.Encoder
 import org.apache.spark.sql.Encoders
 import scala.util.Random
 
-class BloomFilterAggregator(elements:Int, bits:Int)
+class BloomFilterAggregator(elements:Long, bits:Long)
   extends Aggregator[Int, BitSetView, BloomFilter] {
 
   private val numHashFunctions : Int =
